requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.1
fonttools==4.59.2
h11==0.16.0
idna==3.10
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
zipp==3.23.0
